ki-aim-synthetization/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.1
PyYAML==6.0.1
cloudpickle==2.2.1
Flask-SocketIO==5.3.6
packaging>=20,<22
numpy==1.26.4
pandas>=2.1
scikit-learn>=0.24,<2
torch>=2.2
torchvision>=0.9.0,<1
rdt>=1.2.0,<2.0
tqdm>=4.15,<5
gunicorn>=22.0.0
requests==2.32.3
numpy==1.26.4
joblib==1.4.2
synthcity==0.2.11